add day06

+++ go/aoc24/day06/part_1_test.go
@@ -5,8 +5,8 @@ import (
 )
 
 func TestOutOfBoundsShouldBeInside(t *testing.T) {
-	x := 2
-	y := 2
+	x := int16(2)
+	y := int16(2)
 	testField := NewField(10, 5, nil)
 	if OutOfBounds(&testField, &x, &y) {
 		t.Errorf(`Expected %s to match`, "a")
@@ -14,14 +14,14 @@ func TestOutOfBoundsShouldBeInside(t *testing.T) {
 }
 
 func TestOutOfBoundsShouldBeOutside(t *testing.T) {
-	x := -1
-	y := 5
+	x := int16(-1)
+	y := int16(5)
 	testField := NewField(6, 6, nil)
 	if !OutOfBounds(&testField, &x, &y) {
 		t.Errorf(`Expected %s to match`, "a")
 	}
-	x = 1
-	y = -5
+	x = int16(1)
+	y = int16(-5)
 	testField = NewField(6, 6, nil)
 	if !OutOfBounds(&testField, &x, &y) {
 		t.Errorf(`Expected %s to match`, "a")
